refactor: squad moderation to source moderation
