+++ src/workers/notifications/sourcePostModerationRejected.ts
@@ -1,43 +1,32 @@
 import { generateTypedNotificationWorker } from './worker';
-import { SourceMember } from '../../entity';
-import { In } from 'typeorm';
-import { SourceMemberRoles } from '../../roles';
 import { NotificationType } from '../../notifications/common';
 import { NotificationPostModerationContext } from '../../notifications';
-import { SquadPostModerationStatus } from '../../entity/SquadPostModeration';
+import { SquadPostModerationStatus } from '../../entity/sourcePostModeration';
 import { getPostModerationContext } from './utils';
 import { logger } from '../../logger';
 import { TypeORMQueryFailedError } from '../../errors';
 
 const worker =
   generateTypedNotificationWorker<'api.v1.source-post-moderation-submitted'>({
-    subscription: 'api.v1.source-post-moderation-submitted-notification',
+    subscription: 'api.v1.source-post-moderation-rejected-notification',
     handler: async ({ post }, con) => {
-      if (post.status !== SquadPostModerationStatus.Pending) {
+      if (post.status !== SquadPostModerationStatus.Rejected) {
         return;
       }
 
       try {
         const moderationCtx = await getPostModerationContext(con, post);
-        const mods = await con.getRepository(SourceMember).find({
-          select: ['userId'],
-          where: {
-            sourceId: post.sourceId,
-            role: In([SourceMemberRoles.Admin, SourceMemberRoles.Moderator]),
-          },
-        });
-
         const ctx: NotificationPostModerationContext = {
           ...moderationCtx,
-          userIds: mods.map((m) => m.userId),
+          userIds: [post.createdById],
         };
 
-        return [{ type: NotificationType.SquadPostSubmitted, ctx }];
+        return [{ type: NotificationType.SquadPostRejected, ctx }];
       } catch (err) {
         const error = err as TypeORMQueryFailedError;
         if (error?.name !== 'EntityNotFoundError') {
           logger.error(
-            'failed sending notification for squad post moderation submitted',
+            'failed sending notification for squad post moderation rejected',
             err,
           );
         }
